requirements.txt

beautifulsoup4
disnake>=2.7
GitPython==3.1.30
lavalink>=4.0
numpy==1.22.3
Pillow==9.4.0
pymongo==4.3.3
python-dotenv==0.21.1
pytz==2022.7
requests
spotipy>=2.22
youtube_dl>=2021
git+https://github.com/Snipy7374/disnake-ext-paginator.git
install-jdk